Feature/fix richpresence senderid (#126)
Fix the sender ID for rich presence

I need the sender ID so that I can track which player the rich presence
update is for. I thought I could get this from the sender_id, but that's always zero, since updates are coming from the server. I'm now adding it to the payload directly.